+++ src/File.zig
@@ -5,7 +5,7 @@ const File = @This();
 
 /// The name of the file including any directories.
 ///
-/// Note that iterator does not sanitize the file name.
+/// Note that parser does not sanitize the file name.
 /// It may contain "../" or other directory traversal elements.
 /// If you're using the file name to access the file system,
 /// be sure to sanitize it first.